Re-attempt of the project: Rectified errors/bugs
This commit is another attempt of the project. Previous commit had incorrect and incomplete code, mistakes, synax errors and bugs in several questions.

The commit, however, has correct code for each question (or now has correct code for the questions that needed rectification) and emphasizes an inprovement in the understanding of programming as a whole. Given the time period between the date of the previous commit and the date of this commit, it shows how via EDSA, I as an individual, has excelled in coding/programming throughout my time with the academy.